Refactor: 자원관리 용이성을 위해 Folder의 함수들을 LinkSubFolder로 옮김
